requirements.txt
pandas
numpy
scikit-learn
nltk
spacy
pdfminer.six
python-docx
